requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.10.4
google-generativeai==0.7.2
python-multipart==0.0.9
sqlalchemy==2.0.36

